Parser: move tests to different directories
