+++ Cesium.Parser.Tests/ParserTests/FullParserTests.VariableArithmeticTest.verified.txt
